Update apt::source to use puppetlabs-apt 2.0.0+
Since the newest [sensu-puppet](https://github.com/sensu/sensu-puppet) module requires puppetlabs-apt >=2.0.0 it makes sense to me to migrate this to be puppetlabs-apt >=2.0.0 as well.

This is basically the same changes that @bleuchtang made in #45 except there are no merge conflicts as I pulled the latest master and made the changes there.
Closes #45